Bug fixes and cloud database initialization...
Bug fixes:
- major bug fixes in init script and api related to database switch

Database initialization:

Product database is now initialized locally. After running the init script, database is dumped to a file that is added to repo. After every deploy to Heroku, the dump file is used to "restore" the cloud database. This way the cloud db always has the latest version of the product db.