Automatic CI/CD pipeline...
In order to easily automate the deployment of product API, the app is now containerized using Docker. This allows Github Actions to automatically build a new version that will then be pushed to Docker Hub. On the server side there will be a container called Watchtower listening for new versions of container images to Docker Hub. This release of a new image will trigger the server to automatically pull the latest image and start running it.

TLDR: Any commits to 'main' will be automatically deployed to production.